webapp works better

+++ src/main/java/com/timur/tabel/dao/EmployeeDAO.java
@@ -11,6 +11,5 @@ public interface EmployeeDAO {
 
     List<Employee> getAll();
 
-    List<Employee> getByDepartmentId(Long departmentId);
-
+    List<Employee> getByDepartmentId(Integer departmentId);
 }
