+++ src/main/java/com/timur/tabel/entity/Department.java
@@ -5,7 +5,6 @@ import lombok.Data;
 @Data
 public class Department {
 
-    private long id;
+    private int id;
     private String name;
-
 }
