+++ src/main/java/com/timur/tabel/mapper/DepartmentMapper.java
@@ -3,16 +3,16 @@ package com.timur.tabel.mapper;
 import com.timur.tabel.entity.Department;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
 @Component
 public class DepartmentMapper implements RowMapper<Department> {
+
     @Override
     public Department mapRow(ResultSet resultSet, int i) throws SQLException {
         Department department = new Department();
-        department.setId(resultSet.getLong("id"));
+        department.setId(resultSet.getInt("id"));
         department.setName(resultSet.getString("dep_name"));
         return department;
     }
